+++ packages/cryptoplease/lib/data/db/db.dart
@@ -2,6 +2,7 @@ import 'package:drift/drift.dart';
 import 'package:injectable/injectable.dart';
 
 import '../../../../core/transactions/tx_sender.dart';
+import '../../features/activities/module.dart';
 import '../../features/incoming_split_key_payments/module.dart';
 import '../../features/outgoing_direct_payments/module.dart';
 import '../../features/outgoing_split_key_payments/module.dart';
@@ -19,7 +20,7 @@ class OutgoingTransferRows extends Table {
   Set<Column<dynamic>>? get primaryKey => {id};
 }
 
-const int latestVersion = 19;
+const int latestVersion = 20;
 
 const _tables = [
   OutgoingTransferRows,
@@ -27,6 +28,7 @@ const _tables = [
   ODPRows,
   OSKPRows,
   ISKPRows,
+  TransactionRows,
 ];
 
 @lazySingleton
@@ -65,14 +67,20 @@ class MyDatabase extends _$MyDatabase {
           if (from < 17) {
             await m.createTable(iSKPRows);
           }
+
           if (from >= 15 && from < 18) {
             await m.addColumn(oDPRows, oDPRows.reference);
           }
+
           if (from >= 15 && from < 19) {
             await m.addColumn(oDPRows, oDPRows.txFailureReason);
           }
           if (from >= 16 && from < 19) {
             await m.addColumn(oSKPRows, oSKPRows.txFailureReason);
+          }
+
+          if (from < 20) {
+            await m.createTable(transactionRows);
           }
         },
       );