+++ packages/cryptoplease/lib/features/pending_activities/module.dart
@@ -1,1 +1,0 @@
-export 'src/components/pending_activities_list.dart';